requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
google-generativeai
pypdf
requests
python-multipart